requirements.txt
beautifulsoup4
chromadb
html2text
pdfplumber
python-dotenv
requests
